varius fixes (navbar, loan calculator math, stocks)

+++ script3.js
@@ -9,7 +9,7 @@ button.addEventListener("click", function(event) {
 function needed(e){
     e.preventDefault();
     const m = parseInt(monthly.value);
-    const R = parseInt(rate.value)
+    const R = parseFloat(rate.value)
     const p = parseInt(principal.value)
     const r = (R / 12) / 100
     let n = Math.log(m / (m - p * r)) / Math.log(1 + r);
